Fix high-speed SWIO bootloaders
Somehow the assembler (or linker?) of gcc converts the assembler sequence
rcall .+0 followed by ret to rjmp .+0 followed by ret. This distorts the
timing of SWIO bootloaders at the high-speed end of available baud rates,
eg, 576,000 baud at 16 MHz FCPU. This was fixed in the urboot.c source:
https://github.com/stefanrueger/urboot/commit/80d5ea90a9903d3ebea88f1644b302341563f367

This commit finally corrects the generated bootloaders that suffered from
this problem.